Move the existing .paypal.PayPalButtonsFragment to .paypalbuttons.PayPalButtonsFragment.
