+++ Demo/src/main/java/com/paypal/android/ui/paypalbuttons/PayPalButtonsFragment.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.ui.paypal
+package com.paypal.android.ui.paypalbuttons
 
 import android.os.Bundle
 import android.view.LayoutInflater
